+++ src/lib.rs
@@ -55,7 +55,7 @@ pub mod prelude {
         filter::TiledFilter,
         helpers::{
             get_layer_from_map, get_object_from_map, get_tile_from_map, get_tileset_from_map,
-            grid_size_from_map, tile_size, tile_size_from_map, tilemap_type_from_map,
+            grid_size_from_map, tile_size, tilemap_type_from_map,
         },
         image::TiledImage,
         layer::TiledLayer,